switch AdjustedSpace to float for consistent precision in fxlen and update types accordingly

+++ ThrEd4/globals.cpp
@@ -95,7 +95,7 @@ DRAWITEMSTRUCT* DrawItem;             // for owner-draw windows
 POINT           EndPointCross;        // point to draw at the end cross for form select points
 HMENU           FillMenu;             // fill submenu
 
-std::vector<double>* FormAngles;        // angles of a form for satin border fills
+std::vector<float>*  FormAngles;        // angles of a form for satin border fills
 std::vector<POINT>*  FormControlPoints; // form control rectangle in pixel coordinates
 
 HWND                 FormDataSheet = nullptr; // form data sheet
